prevent clearing project root directory or non-project directories

+++ index.js
@@ -31,6 +31,14 @@ task(NAME, DESC, async function (args, hre) {
 
   const outputDirectory = path.resolve(hre.config.paths.root, config.path);
 
+  if (!outputDirectory.startsWith(hre.config.paths.root)) {
+    throw new HardhatPluginError('resolved path must be inside of project directory');
+  }
+
+  if(outputDirectory === hre.config.paths.root) {
+    throw new HardhatPluginError('resolved path must not be root directory');
+  }
+
   if (config.clear) {
     fs.rmdirSync(outputDirectory, { recursive: true });
   }
